fix: ts-build failures with allowJs closes #70
ts-build  assumed it can remove the "outDir" for simpler path calculations. This is not the case when "allowJs" is on, as diagnostics will be emitted saying output overwrites input.

Following this commit, ts-build now leaves "outDir" in-place, letting typecheck properly pass. Each target path is now calculated using the source file path distance from the root src directory.

Behavior should be the same, except it should fail in the allowJs case.